move  ActivationsKey and SemverCompatibility
