+++ src/cargo/core/resolver/context.rs
@@ -1,13 +1,12 @@
 use super::dep_cache::RegistryQueryer;
 use super::errors::ActivateResult;
-use super::types::{ConflictMap, ConflictReason, FeaturesSet, ResolveOpts};
+use super::types::{ActivationsKey, ConflictMap, ConflictReason, FeaturesSet, ResolveOpts};
 use super::RequestedFeatures;
-use crate::core::{Dependency, PackageId, SourceId, Summary};
+use crate::core::{Dependency, PackageId, Summary};
 use crate::util::interning::InternedString;
 use crate::util::Graph;
 use anyhow::format_err;
 use std::collections::HashMap;
-use std::num::NonZeroU64;
 use tracing::debug;
 
 // A `Context` is basically a bunch of local resolution information which is
@@ -39,38 +38,8 @@ pub type ContextAge = usize;
 /// By storing this in a hash map we ensure that there is only one
 /// semver compatible version of each crate.
 /// This all so stores the `ContextAge`.
-pub type ActivationsKey = (InternedString, SourceId, SemverCompatibility);
-
 pub type Activations =
     im_rc::HashMap<ActivationsKey, (Summary, ContextAge), rustc_hash::FxBuildHasher>;
-
-/// A type that represents when cargo treats two Versions as compatible.
-/// Versions `a` and `b` are compatible if their left-most nonzero digit is the
-/// same.
-#[derive(Clone, Copy, Eq, PartialEq, Hash, Debug, PartialOrd, Ord)]
-pub enum SemverCompatibility {
-    Major(NonZeroU64),
-    Minor(NonZeroU64),
-    Patch(u64),
-}
-
-impl From<&semver::Version> for SemverCompatibility {
-    fn from(ver: &semver::Version) -> Self {
-        if let Some(m) = NonZeroU64::new(ver.major) {
-            return SemverCompatibility::Major(m);
-        }
-        if let Some(m) = NonZeroU64::new(ver.minor) {
-            return SemverCompatibility::Minor(m);
-        }
-        SemverCompatibility::Patch(ver.patch)
-    }
-}
-
-impl PackageId {
-    pub fn as_activations_key(self) -> ActivationsKey {
-        (self.name(), self.source_id(), self.version().into())
-    }
-}
 
 impl ResolverContext {
     pub fn new() -> ResolverContext {
